Create number group variable to facilitate merging of numbers and names in long format
Previously, names and numbers were merged based on interview ID and number type. This resulted in duplicates. Now, an additional variable is created to serve as another key that facilitates merging. To create that variable, the `create_preload_files` now has another argument that takes the formula for creating this key as a quosure.